src/demo-utils/pyproject.toml
[project]
name = "demo-utils"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.14",
    "azure-ai-projects>=1.0.0b12",
    "azure-cosmos>=4.9.0",
    "azure-data-tables>=12.7.0",
    "azure-identity>=1.23.0",
    "azure-search-documents>=11.5.3",
    "azure-servicebus>=7.12.0",
    "azure-storage-blob>=12.23.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.1.1",
]
